Clarify default behaviour when channel is not set
Got a small surprise when the bot started posting to #general, it will
be useful to clarify this to new users. I also avoided specifying a
channel because I didn't want to create a spam channel, but it's still
quite useful because it will still mention the user. Arguably more
useful because it encourages collaboration by sharing information.