Fix download link
The current link goes to a page that says
>No direct downloads selected for this package.

Unless you go through the closed issues, it's not clear that there are
packages already available for download. I believe the `Downloads`
section on bintray should be populated, but perhaps this can help people
find the download link in the meantime.