Traverse up the material dependency tree for someone to report to.
An exception was being thrown if the plugin encountered a material
that was a dependency to another pipeline.

This change will traverse up the dependency tree until it finds a
material that is not a dependency on another pipeline. It will also
process all modifications to a material revision, reporting to all users
that the build status might concern.

This change will also change the status message formatting a bit, make
it more email friendly. Please take an extra look at this change.

Fixes: https://github.com/gmazzo/gocd-build-watcher-plugin/issues/7